refactor: 문자열 포맷팅 및 패키지 이동
