+++ src/main/java/roomescape/config/WebMvcConfig.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.handler.LoginUserArgumentResolver;
+import roomescape.infrastructure.LoginUserArgumentResolver;
 import roomescape.infrastructure.CheckAdminInterceptor;
 
 @Configuration
@@ -19,7 +19,6 @@ public class WebMvcConfig implements WebMvcConfigurer {
         this.loginUserArgumentResolver = loginUserArgumentResolver;
         this.checkAdminInterceptor = checkAdminInterceptor;
     }
-
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
