+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,21 +1,20 @@
 package roomescape.controller;
 
 import java.net.URI;
-import java.time.LocalDate;
 import java.util.List;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
 import roomescape.service.ReservationService;
 import roomescape.service.dto.request.LoginUser;
+import roomescape.service.dto.request.ReservationConditionRequest;
 import roomescape.service.dto.request.ReservationRequest;
 import roomescape.service.dto.response.ReservationResponse;
 
@@ -49,13 +48,9 @@ public class ReservationController {
     }
 
     @GetMapping(params = {"themeId", "memberId", "dateFrom", "dateTo"})
-    public ResponseEntity<List<ReservationResponse>> getReservationsOf(@RequestParam Long themeId,
-                                                                       @RequestParam Long memberId, @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
-                                                                     LocalDate dateFrom,
-                                                                       @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
-                                                                     LocalDate dateTo
-    ) {
-        return ResponseEntity.ok(reservationService.findAllReservationsOf(themeId, memberId, dateFrom, dateTo));
+    public ResponseEntity<List<ReservationResponse>> getReservationsOf(@ModelAttribute ReservationConditionRequest request
+                                                                       ) {
+        return ResponseEntity.ok(reservationService.findAllReservationsByCondition(request));
     }
 
     @DeleteMapping("/{id}")
