+++ src/main/java/roomescape/domain/MemberRepository.java
@@ -6,7 +6,10 @@ import java.util.Optional;
 public interface MemberRepository {
 
     Member save(Member member);
+
     Optional<Member> findById(Long id);
+
     Optional<Member> findByEmail(String email);
+
     List<Member> findAllMembers();
 }
