+++ src/main/resources/static/js/reservation-with-member.js
@@ -192,10 +192,6 @@ function applyFilter(event) {
   const dateFrom = document.getElementById('date-from').value;
   const dateTo = document.getElementById('date-to').value;
 
-  /*
-  TODO: [6단계] 예약 검색 - 조건에 따른 예약 조회 API 호출
-        요청 포맷에 맞게 설정
-  */
   fetch(`/reservations?themeId=${themeId}&memberId=${memberId}&dateFrom=${dateFrom}&dateTo=${dateTo}`, { // 예약 검색 API 호출
     method: 'GET',
     headers: {
